Finished testing checks in sfdb_checks
- Unit testing of check method was direly needed
- Refactoring of index shift between human-readable line index and machine-style entry-index was needed
- Added some ToDos that came up along the way.
- Made output of checks more readily human readable by moving the conversion of entries to entry-strings into the checks
- Made __main__ more readable by applying the test-pattern also to comparison tests

Overall this commit finishes unit testing for one half of the sfdb_checks library. The log-functions still require tests to be written for them. In the context of writing these tests, the general design pattern for the check functions to have the check happen in one function and its logging in another was applied to the comparison test as well and that moved over to the __main__ function that needed adjusting accordingly. Introduction of "INDEX_SHIFT" was needed as first step to more easily gap the bridge between an entry index (start at 0, start at the first entry of the SFDB) and a line-index of the SFDB file (start at 1, start at the first lines of the SFDB file including the header).