Major overhaul of logging
- Logging format was ugly and needed to be beautified
- Logging code was ugly and needed to be standardized and simplified
- Tests needed to be adjusted accordingly
- Definition of Column objects needed to be adjusted accordingly
- Format of sfdb_schemas.json needed to be adjusted

This commit constitutes a major overhaul of the way logging is done, mostly because the format of the logs is nicer as well as the code being more pythonic. In the wake of this change, many adjustments were made. Column objects defined in sql_table_schema.py now include the name of the column itself. This warranted a change in sfdb_schema.json which was not as easy to read as I liked. All of this affected various tests and even some sfdb_checks themselves which thusly had some of their code moved out into smaller functions in order to simplify them.