Finished testing argparser.py
- argparser wasn't yet finished
- While writing unittest the need for refactoring and adding to argparse became apparent
- created custom ArgumentParser to avoid argparse's own error-handling
- added additional checks of arguments to be able to give out more helpful error messages.

Reasons are as written above. As part of the simplification process sfdb's are now read in during the argument parsing step and turned into SFDBContainer objects, which are then set as attribute of args. This allows for much earlier catching of faulty inputs. In the same vein sfdbs now have an attribute to "filepath" to contain the file they were read in from, if they were read from a file. That was in part necessary as logging relied on the name of the SFDB file to create the name of the log file.