Refactored variable names with "line" and "entry"
- the sfdb-module implicitly enforces the separate ideas of "line" and "entry"
- lines are lines in an SFDB file (start at SFDB file line-index 0)
- entries are SFDB database entries (start at SFDB File line-index 5)
- This was inconsistently applied throughout the program, which required fixing
- Minor formatting tweaks

This commit fixes the issue by adjusting the variable names. It further enforces it by ensuring that sfdb_check methods return lists of **entries**, not lines (exception being check_content_against_regex which works directly on lines). This required some adjustments on the side of the unit-tests as well.

Sidenote: I received feedback regarding some of the table layout of the log and incorporated minor tweaks to it accordingly.