Fixed Bug in argparser when requesting SFDB
When in request mode the program would upon receiving an SFDB filepath put the filepath into the args object, not the SFDB itself. This was ammended by implementing "request_sfdb", which required unittesting.
In the wake of the unittesting the need to implement __eq__, __hash__ and a more detailed version of __str__ became obvious. The old version of __repr__ became the new __str__ and __repr__ received a more detailed output, now including the entries.